moved restaurants app to apof dir
