+++ src/apof/restaurants/apps.py
